pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autonoma"
version = "0.1.0"
description = "Autonoma — Open-source AI agent platform. Build, deploy, and run digital employees."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "websockets>=13.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.scripts]
autonoma = "autonoma.main:cli_entry"

[tool.setuptools.packages.find]
include = ["autonoma*"]
